Mascaras

+++ app/src/main/java/com/pliniodev/finanassimples_controlefinanceiropessoal/view/activity/RegisterActivity.kt
@@ -19,7 +19,7 @@ import java.util.*
 class RegisterActivity : AppCompatActivity(), View.OnClickListener {
 
     /**
-     * Activitie onde são feitos os registros de despesas e receitas do usuário
+     * Activity onde são feitos os registros de despesas e receitas do usuário
      */
 
     private lateinit var binding: ActivityRegisterBinding
@@ -27,8 +27,6 @@ class RegisterActivity : AppCompatActivity(), View.OnClickListener {
     private var mTransactionId: Int = 0
     private var mMes = 0
     private var mCalendar = Calendar.getInstance()
-
-
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -131,8 +129,7 @@ class RegisterActivity : AppCompatActivity(), View.OnClickListener {
             val name = binding.editName.text.toString()
             val description = binding.editDescription.text.toString()
 
-//            val df = DecimalFormat("###,###.##")
-            val price = binding.editPrice.text.toString().toDouble()
+            val price = binding.editPrice.text.toString()
 
             val category = binding.editCategory.text.toString()
             val dueDate = binding.textEditDate.text.toString()
@@ -140,8 +137,18 @@ class RegisterActivity : AppCompatActivity(), View.OnClickListener {
             val paidOut = binding.switchPay.isChecked
             val observation = binding.observationExtra.text.toString()
 
-            mViewModel.save(mTransactionId, transactionType, name, description, price, category,
-                    dueDate, month, paidOut, observation)
+            mViewModel.save(
+                mTransactionId,
+                transactionType,
+                name,
+                description,
+                price,
+                category,
+                dueDate,
+                month,
+                paidOut,
+                observation
+            )
         }
     }
